replace httpapi with websocketapi
